Refactor: rename discrete_representation to discrete
