+++ coverage.py
@@ -17,7 +17,7 @@ untracked_modules = ['volmdlr/templates.py',
                      'volmdlr/bspline_compiled.py',
                      "volmdlr/bspline_fitting.py",
                      "volmdlr/bspline_evaluators.py",
-                     "volmdlr/discrete_representation_compiled.py",
+                     "volmdlr/discrete_compiled.py",
                      "volmdlr/stl.py",
                      'nurbs/core.py',
                      'nurbs/helpers.py',
