+++ scripts/discrete_representation/pixelization/pixelization_filling.py
@@ -4,7 +4,7 @@ Showcase of filling a pixelization.
 
 from volmdlr import Point2D
 from volmdlr.core import EdgeStyle
-from volmdlr.discrete_representation import MatrixBasedPixelization
+from volmdlr.discrete import MatrixBasedPixelization
 from volmdlr.wires import ClosedPolygon2D
 
 PIXEL_SIZE = 0.05
