+++ scripts/discrete_representation/voxelization/compare_memory_usage.py
@@ -6,7 +6,7 @@ import time
 
 import volmdlr
 from pympler.asizeof import asizeof
-from volmdlr.discrete_representation import MatrixBasedVoxelization, OctreeBasedVoxelization, PointBasedVoxelization
+from volmdlr.discrete import MatrixBasedVoxelization, OctreeBasedVoxelization, PointBasedVoxelization
 from volmdlr.step import Step
 
 STEP_MODEL_FILE_PATH = "../../step/tore1.step"
