+++ scripts/discrete_representation/voxelization/interference/interference_volume_showcase.py
@@ -6,7 +6,7 @@ import time
 
 import volmdlr
 from volmdlr.core import VolumeModel
-from volmdlr.discrete_representation import PointBasedVoxelization
+from volmdlr.discrete import PointBasedVoxelization
 from volmdlr.step import Step
 
 STEP_MODEL_FILE_PATH = "../../../step/tore1.step"
