+++ scripts/discrete_representation/voxelization/octree/non_homogeneous_showcase.py
@@ -5,7 +5,7 @@ import random
 
 import volmdlr
 from volmdlr.core import VolumeModel
-from volmdlr.discrete_representation import MatrixBasedVoxelization
+from volmdlr.discrete import MatrixBasedVoxelization
 from volmdlr.primitives3d import Cylinder
 
 VOXEL_SIZE = 0.01
