+++ scripts/discrete_representation/voxelization/stl_file_voxelization.py
@@ -1,7 +1,7 @@
 """
 Example of voxelization from an STL file.
 """
-from volmdlr.discrete_representation import MatrixBasedVoxelization
+from volmdlr.discrete import MatrixBasedVoxelization
 from volmdlr.display import Mesh3D
 from volmdlr.core import VolumeModel
