+++ scripts/discrete_representation/voxelization/volume_model_point_based_voxelization.py
@@ -5,7 +5,7 @@ In PointBasedVoxelization, voxels are stored as a Set[Tuple[float, float, float]
 """
 import volmdlr
 from volmdlr.core import VolumeModel
-from volmdlr.discrete_representation import PointBasedVoxelization
+from volmdlr.discrete import PointBasedVoxelization
 from volmdlr.primitives3d import Cylinder, Sphere
 
 VOXEL_SIZE = 0.01
