+++ tests/discrete_representation/test_discrete_representation.py
@@ -4,7 +4,7 @@ Unit testing of 'DiscreteRepresentation' class.
 import math
 import unittest
 
-from volmdlr.discrete_representation import DECIMALS, DiscreteRepresentation
+from volmdlr.discrete import DECIMALS, DiscreteRepresentation
 
 
 class TestDiscreteRepresentation(unittest.TestCase):
