+++ tests/discrete_representation/test_matrix_based_voxelization.py
@@ -6,7 +6,7 @@ import unittest
 
 import volmdlr
 from volmdlr.core import BoundingBox, VolumeModel
-from volmdlr.discrete_representation import MatrixBasedVoxelization
+from volmdlr.discrete import MatrixBasedVoxelization
 from volmdlr.primitives3d import Block, Cylinder, Sphere
 from volmdlr.shells import ClosedTriangleShell3D, DisplayTriangleShell3D
