+++ tests/discrete_representation/test_octree_based_voxelization.py
@@ -5,7 +5,7 @@ import unittest
 
 import volmdlr
 from volmdlr.core import BoundingBox, VolumeModel
-from volmdlr.discrete_representation import OctreeBasedVoxelization
+from volmdlr.discrete import OctreeBasedVoxelization
 from volmdlr.primitives3d import Block, Cylinder, Sphere
 from volmdlr.shells import ClosedTriangleShell3D
