+++ tests/discrete_representation/test_point_based_voxelization.py
@@ -7,7 +7,7 @@ import unittest
 
 import volmdlr
 from volmdlr.core import BoundingBox, VolumeModel
-from volmdlr.discrete_representation import PointBasedVoxelization
+from volmdlr.discrete import PointBasedVoxelization
 from volmdlr.primitives3d import Block, Cylinder, Sphere
 from volmdlr.shells import ClosedTriangleShell3D, DisplayTriangleShell3D
