+++ volmdlr/core.py
@@ -25,7 +25,7 @@ import dessia_common.files as dcf
 import volmdlr
 import volmdlr.templates
 from volmdlr.core_compiled import bbox_is_intersecting
-from volmdlr.discrete_representation_compiled import triangle_intersects_voxel
+from volmdlr.discrete_compiled import triangle_intersects_voxel
 from volmdlr.utils.step_writer import product_writer, geometric_context_writer, assembly_definition_writer, \
     STEP_HEADER, STEP_FOOTER, step_ids_to_str
 from volmdlr.geometry import get_transfer_matrix_from_basis
